Fix bug in family generator caused by species changes
